requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-multiselectfield==1.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==25.0
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
sqlparse==0.5.3
stripe==12.3.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
celery==5.4.0
django-redis==5.4.0
redis==5.2.1
psycopg2-binary
